+++ tests/FreeElephants/TSxUnit/Mock/WindowBuilderTest.ts
@@ -4,7 +4,8 @@ namespace FreeElephants.TSxUnit.Mock {
     export class WindowBuilderTest extends TestCase {
 
         public testStubWindowFromUrl() {
-            let window: Window = WindowBuilder.stubWindowFromUrl("http://example.com");
+            let builder: WindowBuilder = new WindowBuilder();
+            let window: Window = builder.setLocation("http://example.com", true).getMock();
             this.assertEquals("example.com", window.location.host);
             this.assertEquals("Example Domain", window.document.title);
         }